extract createMenu function in util directory
